requirements.txt
flask
redis
psycopg2-binary
python-dotenv
pytest
gunicorn
opencensus-ext-azure
opencensus-ext-flask
